Reconfigure httpd container, for real certs
After building an httpd container, the certificates will be present in the image are for HTTPS with a self-signed certificates. If real certificates are required, the real certificates should be placed in the /usr/local/docker-config/httpd/cert directory on the host machine, where it will be volume mounted from.